Install Chrome

+++ install.sh
@@ -11,14 +11,7 @@ if test ! -d /usr/local/go; then
   tar xvf go1*.tar.gz
   sudo chown -R root:root ./go
   sudo mv go /usr/local
-fi
-
-# Oh-my-zsh
-if [ ! -d "$HOME/.oh-my-zsh" ]; then
-  printf "Installing ohmyzsh\n"
-	sh -c "$(curl -fsSL https://raw.github.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
-	git clone https://github.com/zsh-users/zsh-autosuggestions "${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions"
-	git clone https://github.com/zsh-users/zsh-syntax-highlighting.git "${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting"
+  rm go1*.tar.gz
 fi
 
 # Fuzzy search
@@ -33,12 +26,19 @@ if test ! -f /usr/bin/docker; then
   printf "Installing Docker\n"
   sudo apt install apt-transport-https ca-certificates curl software-properties-common -y
   curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
-  sudo add-apt-repository "deb [arch=amd64] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable"
+  sudo add-apt-repository "deb [arch=amd64] https://download.docker.com/linux/ubuntu eoan stable"
   sudo apt update
-  sudo apt-get install -y docker-ce docker-ce-cli containerd.io docker-ce docker-compose
+  sudo apt-get install -y docker-ce docker-ce-cli containerd.io docker-ce
   sudo groupadd docker
   sudo usermod -aG docker "${USER}"
   newgrp docker
+fi
+
+# Chrome
+if ! dpkg -s google-chrome-stable; then
+  wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
+  sudo dpkg -i google-chrome-stable_current_amd64.deb
+  rm google-chrome-stable_current_amd64.deb
 fi
 
 # Spotify
@@ -67,9 +67,11 @@ if ! dpkg -s slack-desktop; then
 fi
 
 # Kubectl
-curl -LO https://storage.googleapis.com/kubernetes-release/release/`curl -s https://storage.googleapis.com/kubernetes-release/release/stable.txt`/bin/linux/amd64/kubectl
-chmod +x ./kubectl
-sudo mv ./kubectl /usr/local/bin/kubectl
+if test ! -f /usr/local/bin/kubectl; then
+  curl -LO https://storage.googleapis.com/kubernetes-release/release/`curl -s https://storage.googleapis.com/kubernetes-release/release/stable.txt`/bin/linux/amd64/kubectl
+  chmod +x ./kubectl
+  sudo mv ./kubectl /usr/local/bin/kubectl
+fi
 
 # Go libraries
 GO111MODULE="on" go get sigs.k8s.io/kind@v0.7.0
@@ -81,12 +83,28 @@ GO111MODULE="on" go get github.com/giantswarm/devctl
 # IntelliJ
 if ! ls /opt/idea-* 1> /dev/null 2>&1; then
   printf "Installing IntelliJ\n"
-  curl -O https://download.jetbrains.com/idea/ideaIU-2020.1.tar.gz
-  sudo tar -xzf ideaIU.tar.gz -C /opt
+  wget https://download.jetbrains.com/idea/ideaIU-2020.1.tar.gz
+  sudo tar -xzf ideaIU*.tar.gz -C /opt
   rm ideaIU*.tar.gz
 fi
 
 # SDKMan
 if test ! -f /home/jose/.sdkman/src/sdkman-main.sh; then
   curl -s "https://get.sdkman.io" | bash
+fi
+
+# Keybase
+if ! dpkg -s keybase; then
+  curl --remote-name https://prerelease.keybase.io/keybase_amd64.deb
+  sudo apt install -y ./keybase_amd64.deb
+  rm keybase_amd64.deb
+  #run_keybase
+fi
+
+# Oh-my-zsh
+if [ ! -d "$HOME/.oh-my-zsh" ]; then
+  printf "Installing ohmyzsh\n"
+  git clone https://github.com/zsh-users/zsh-autosuggestions "${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions"
+  git clone https://github.com/zsh-users/zsh-syntax-highlighting.git "${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting"
+  sh -c "$(curl -fsSL https://raw.github.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
 fi
